notification-bridge/pyproject.toml
[project]
name = "notification-bridge"
version = "0.1.0"
description = "Cross-platform desktop notification bridge - forwards notifications to Central Context API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
linux = ["dbus-fast>=4.0.0"]
windows = [
    "winrt-runtime>=3.0.0",
    "winrt-Windows.UI.Notifications>=3.0.0",
]

[project.scripts]
notification-bridge = "notification_bridge.main:main"

[dependency-groups]
dev = [
    "dbus-fast>=4.0.0",  # For Linux development
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
